+++ scripts/weekly-tileset-generation.py
@@ -6,7 +6,7 @@ from get_frame_list_from_cmr import get_frames_for_direction
 
 
 def publish_mosaic(job: sdk.Job) -> None:
-    source = f's3://hyp3-opera-disp-sandbox-contentbucket-ibxz8lcpdo59/{job.job_id}/tms/'
+    source = f's3://hyp3-edc-prod-contentbucket-1fv14ed36ifj6/{job.job_id}/tms/'
     target = f's3://asf-services-web-content-prod/main/{job.name}/'
     subprocess.run(['aws', '--profile', 'edc-prod', 's3', 'sync', source, target, '--delete'], check=True)
 
@@ -28,11 +28,7 @@ def main():
     ascending_frames = get_frames_for_direction('ASCENDING')
     descending_frames = get_frames_for_direction('DESCENDING')
 
-    # frame 21518 has disjoint temporal coverage
-    if 21518 in ascending_frames:
-        ascending_frames.remove(21518)
-
-    HyP3 = sdk.HyP3('https://hyp3-opera-disp-sandbox.asf.alaska.edu/')
+    HyP3 = sdk.HyP3('https://hyp3-api.asf.alaska.edu/')
 
     prepared_jobs = [
         build_job('desc/disp', 'displacement', descending_frames),
@@ -43,12 +39,12 @@ def main():
 
     jobs = HyP3.submit_prepared_jobs(prepared_jobs)
     for job in jobs:
-        print(f'https://hyp3-opera-disp-sandbox.asf.alaska.edu/jobs/{job.job_id}')
+        print(f'https://hyp3-api.asf.alaska.edu/jobs/{job.job_id}')
     jobs = HyP3.watch(jobs, timeout=21600, interval=120)
 
     for job in jobs:
         response = requests.get(
-            f'https://hyp3-opera-disp-sandbox-contentbucket-ibxz8lcpdo59.s3.us-west-2.amazonaws.com/{job.job_id}/tms/openlayers.html'
+            f'https://d3gm2hf49xd6jj.cloudfront.net/{job.job_id}/tms/openlayers.html'
         )
         response.raise_for_status()